mysite/requirements.txt
asgiref==3.8.1
Django==4.1
django-environ==0.11.2
exceptiongroup==1.2.0
Faker==23.1.0
flake8==7.0.0
gunicorn==21.2.0
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
packaging==24.0
pillow==10.2.0
pluggy==1.4.0
pycodestyle==2.11.1
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.10.0
django-multiupload==0.6.1
cloudinary==1.40.0
social-auth-app-django==5.4.1
